+++ packages/plugin-table/web/src/defaults.ts
@@ -1,4 +1,4 @@
-import { ThemeGeneratorFunction, NO_WRAP } from 'wix-rich-content-common';
+import { ThemeGeneratorFunction } from 'wix-rich-content-common';
 import { COLORS } from './consts';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -14,7 +14,5 @@ export const theme: ThemeGeneratorFunction = ({ colors, utils }) => {
 };
 
 export const DEFAULTS = Object.freeze({
-  config: {
-    textWrap: NO_WRAP,
-  },
+  config: {},
 });