+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -13,6 +13,7 @@ import {
   CustomAnchorScroll,
   Link_Rel,
   AvailableExperiments,
+  LinkPreviewData,
 } from 'wix-rich-content-common';
 import { EditorState, EditorProps } from 'draft-js';
 import { PreviewConfig } from 'wix-rich-content-preview';
@@ -76,6 +77,7 @@ export interface RicosViewerProps extends RicosProps, Pick<RichContentViewerProp
   preview?: PreviewConfig;
   seoSettings?: boolean | SEOSettings;
   textSelectionToolbar?: boolean;
+  linkPreviewPopoverFetchData?: (url: string) => Promise<LinkPreviewData>;
   /* Changes to this interface should also be reflected in the API docs */
 }