+++ packages/ricos-content/web/src/converters/draft/toDraft/toDraft.ts
@@ -20,7 +20,8 @@ import {
 import preprocess from './preprocess';
 
 const convert = (ricosContent: RichContent): DraftContent => {
-  const { nodes, docStyle } = RichContent.toJSON(RichContent.fromJSON(ricosContent)) as RichContent; // using toJSON to remove undefined fields
+  const { nodes, docStyle } = ricosContent;
+
   const draftContent: DraftContent = {
     blocks: [],
     entityMap: {},