+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -24,16 +24,12 @@ import {
   EditorEventsContext,
   EditorEvents,
 } from 'wix-rich-content-editor-common/libs/EditorEventsContext';
-import {
-  ToolbarType,
-  Version,
-  RicosTranslate,
-  EditorCommands,
-  getLangDir,
-} from 'wix-rich-content-common';
+import { ToolbarType, Version, RicosTranslate, getLangDir } from 'wix-rich-content-common';
 import { emptyDraftContent, getEditorContentSummary } from 'wix-rich-content-editor-common';
 import englishResources from 'wix-rich-content-common/dist/statics/locale/messages_en.json';
 import { TextFormattingToolbarType } from './toolbars/TextFormattingToolbar';
+import { getBiFunctions } from './toolbars/utils/biUtils';
+import { isLinkToolbarOpen } from './toolbars/utils/toolbarsUtils';
 
 // eslint-disable-next-line
 const PUBLISH_DEPRECATION_WARNING_v9 = `Please provide the postId via RicosEditor biSettings prop and use one of editorRef.publish() or editorEvents.publish() APIs for publishing.
@@ -46,7 +42,7 @@ interface State {
   remountKey: boolean;
   editorState?: EditorState;
   initialContentChanged: boolean;
-  activeEditor?: RichContentEditor | null;
+  activeEditor: RichContentEditor | null;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   tiptapEditorModule: Record<string, any> | null;
   tiptapToolbar: unknown;
@@ -127,7 +123,6 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
       isMobile ? ToolbarType.MOBILE : useStaticTextToolbar ? ToolbarType.STATIC : ToolbarType.INLINE
     );
     this.props.editorEvents?.subscribe(EditorEvents.RICOS_PUBLISH, this.onPublish);
-    this.setState({ activeEditor: this.editor });
   }
 
   loadEditor() {
@@ -322,68 +317,23 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
     );
   }
 
-  isLinkToolbarOpenOnMobile() {
-    const { activeEditor } = this.state;
-    const { isMobile } = this.props;
-    if (this.useNewFormattingToolbar && activeEditor && isMobile) {
-      const editorCommands: EditorCommands = activeEditor.getEditorCommands();
-      return editorCommands.getSelection().getIsCollapsed && editorCommands.hasLinkInSelection();
-    } else {
-      return false;
-    }
-  }
-
-  getBiFunctions() {
-    const { _rcProps: { helpers } = {} } = this.props;
-    const onInlineToolbarOpen = toolbarType =>
-      helpers?.onInlineToolbarOpen?.({
-        toolbarType,
-        version: Version.currentVersion,
-      });
-    const onToolbarButtonClick = (
-      name: string,
-      toolbarType: ToolbarType,
-      value?: string | boolean,
-      pluginId?: string
-    ) => {
-      helpers?.onToolbarButtonClick?.({
-        buttonName: name,
-        pluginId,
-        type: toolbarType,
-        value: value === undefined ? undefined : typeof value === 'boolean' ? `${!value}` : value,
-        version: Version.currentVersion,
-      });
-      if (pluginId && value && typeof value === 'string') {
-        helpers?.onPluginAdd?.(pluginId, 'FormattingToolbar', Version.currentVersion);
-        helpers?.onPluginAddSuccess?.(pluginId, 'FormattingToolbar', value, Version.currentVersion);
-      }
-    };
-    return { onInlineToolbarOpen, onToolbarButtonClick };
-  }
-
-  renderDraftEditor() {
-    const { TextFormattingToolbar, StaticToolbar, remountKey, activeEditor } = this.state;
+  renderNewToolbars() {
+    const { TextFormattingToolbar, StaticToolbar, activeEditor } = this.state;
     const {
       isMobile,
       theme,
       locale,
+      _rcProps: { helpers } = {},
       toolbarSettings: { getToolbarSettings = () => [] } = {},
       plugins,
       linkPanelSettings,
       linkSettings,
+      experiments,
     } = this.props;
-    const child =
-      this.props.children && shouldRenderChild('RichContentEditor', this.props.children) ? (
-        this.props.children
-      ) : (
-        <RichContentEditor />
-      );
-
     const activeEditorIsTableCell = activeEditor?.isInnerRCERenderedInTable();
     const textToolbarType = StaticToolbar ? 'static' : null;
-    const hideFormattingToolbar = this.isLinkToolbarOpenOnMobile();
-    const biFunctions = this.getBiFunctions();
-
+    const hideFormattingToolbar = isMobile && isLinkToolbarOpen(activeEditor);
+    const biFunctions = helpers && getBiFunctions(helpers);
     const toolbarsProps = {
       textToolbarType,
       isMobile,
@@ -393,24 +343,46 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
       linkPanelSettings,
       linkSettings,
       ...biFunctions,
+      experiments,
     };
     const baseStyles = { flex: 'none' };
     const baseMobileStyles = { ...baseStyles, position: 'sticky', top: 0, zIndex: 9 };
     return (
+      !activeEditorIsTableCell &&
+      activeEditor && (
+        <div
+          data-hook={'ricos-editor-toolbars'}
+          style={isMobile ? baseMobileStyles : baseStyles}
+          dir={getLangDir(locale)}
+        >
+          {!hideFormattingToolbar && TextFormattingToolbar && (
+            <TextFormattingToolbar activeEditor={activeEditor} {...toolbarsProps} />
+          )}
+          <LinkToolbar activeEditor={activeEditor} {...toolbarsProps} />
+        </div>
+      )
+    );
+  }
+
+  renderToolbars() {
+    const { StaticToolbar } = this.state;
+    return this.useNewFormattingToolbar
+      ? this.renderNewToolbars()
+      : this.renderToolbarPortal(StaticToolbar);
+  }
+
+  renderDraftEditor() {
+    const { remountKey } = this.state;
+    const child =
+      this.props.children && shouldRenderChild('RichContentEditor', this.props.children) ? (
+        this.props.children
+      ) : (
+        <RichContentEditor />
+      );
+
+    return (
       <Fragment key={`${remountKey}`}>
-        {!this.useNewFormattingToolbar && this.renderToolbarPortal(StaticToolbar)}
-        {this.useNewFormattingToolbar && !activeEditorIsTableCell && activeEditor && (
-          <div
-            data-hook={'ricos-editor-toolbars'}
-            style={isMobile ? baseMobileStyles : baseStyles}
-            dir={getLangDir(locale)}
-          >
-            {!hideFormattingToolbar && TextFormattingToolbar && (
-              <TextFormattingToolbar activeEditor={activeEditor} {...toolbarsProps} />
-            )}
-            <LinkToolbar activeEditor={activeEditor} {...toolbarsProps} />
-          </div>
-        )}
+        {this.renderToolbars()}
         {this.renderRicosEngine(child, {
           onChange: this.onChange(child.props.onChange),
           ref: this.setEditorRef,