+++ packages/ricos-editor/web/src/utils/editorCommandsTestsUtil.ts
@@ -101,6 +101,7 @@ const giphy = {
   type: RICOS_GIPHY_TYPE,
   nodeType: Node_Type.GIF,
   data1: GIFData.fromJSON({
+    containerData: {},
     original: {
       gif: 'https://media2.giphy.com/media/wZcio4ufvvdde/giphy.gif',
       mp4: 'https://media2.giphy.com/media/wZcio4ufvvdde/giphy.mp4',
@@ -115,6 +116,7 @@ const giphy = {
     width: 500,
   }),
   data2: GIFData.fromJSON({
+    containerData: {},
     original: {
       gif: 'https://media2.giphy.com/media/wZcio4ufvvdde/giphy.gif',
       mp4: 'https://media2.giphy.com/media/wZcio4ufvvdde/giphy.mp4',
@@ -179,8 +181,8 @@ const giphy = {
 const html = {
   type: RICOS_HTML_TYPE,
   nodeType: Node_Type.HTML,
-  data1: HTMLData.fromJSON({ url: 'www.wix.com' }),
-  data2: HTMLData.fromJSON({ url: 'www.sport5.co.il' }),
+  data1: HTMLData.fromJSON({ url: 'www.wix.com', containerData: {} }),
+  data2: HTMLData.fromJSON({ url: 'www.sport5.co.il', containerData: {} }),
   expectedData1: {
     config: {
       alignment: 'center',
@@ -205,6 +207,7 @@ const gallery = {
   type: RICOS_GALLERY_TYPE,
   nodeType: Node_Type.GALLERY,
   data1: GalleryData.fromJSON({
+    containerData: {},
     items: [
       {
         image: {
@@ -260,6 +263,7 @@ const gallery = {
     },
   }),
   data2: GalleryData.fromJSON({
+    containerData: {},
     items: [
       {
         image: {
@@ -451,6 +455,7 @@ const image = {
   type: RICOS_IMAGE_TYPE,
   nodeType: Node_Type.IMAGE,
   data1: ImageData.fromJSON({
+    containerData: {},
     image: {
       src: { custom: '8bb438_1b73a6b067b24175bd087e86613bd00c.jpg' },
       width: 1920,