+++ packages/toolbars-new/web/src/Toolbar/buttonsListCreator.tsx
@@ -34,7 +34,8 @@ export const createButtonsList = (
   plugins,
   linkPanelData,
   colorPickerData,
-  headingsData
+  headingsData,
+  experiments
 ) => {
   const buttonsList = [];
   formattingButtonsKeys.forEach((buttonKey, index) => {
@@ -47,7 +48,7 @@ export const createButtonsList = (
     handleButtonPlugin(buttonsList, index);
     handleButtonLabel(buttonsList, index, editorCommands, t);
     handleButtonArrow(buttonsList, index);
-    handleButtonOnClick(buttonsList, index, editorCommands, linkPanelData);
+    handleButtonOnClick(buttonsList, index, editorCommands, linkPanelData, experiments);
     handleButtonIsActive(buttonsList, index, editorCommands);
     handleButtonIsDisabled(buttonsList, index, editorCommands);
     handleButtonModal(buttonsList, index, editorCommands, linkPanelData, headingsData, t);
@@ -338,7 +339,13 @@ const handleButtonIsActive = (buttonsList, index, editorCommands: editorCommands
   }
 };
 
-const handleButtonOnClick = (buttonsList, index, editorCommands: editorCommands, linkPanelData) => {
+const handleButtonOnClick = (
+  buttonsList,
+  index,
+  editorCommands: editorCommands,
+  linkPanelData,
+  experiments
+) => {
   const buttonName = buttonsList[index].name;
   if (Object.keys(inlineStyleButtons).includes(buttonName)) {
     buttonsList[index].onClick = () =>
@@ -365,7 +372,7 @@ const handleButtonOnClick = (buttonsList, index, editorCommands: editorCommands,
     buttonsList[index].onClick = () => editorCommands.redo();
   } else if (buttonName === 'goToLink') {
     buttonsList[index].onClick = event =>
-      goToLink(event, editorCommands.getLinkDataInSelection(), linkPanelData);
+      goToLink(event, editorCommands.getLinkDataInSelection(), linkPanelData, experiments);
   } else if (buttonName === 'AddPlugin') {
     buttonsList[index].onClick = () => {
       const addPluginButton = document.querySelector(
@@ -535,7 +542,7 @@ const updateDynamicStyles = (type, editorCommands: editorCommands, buttonName) =
   editorCommands.insertDecoration(decorationButtons[buttonName], { ...data });
 };
 
-const goToLink = (event, linkData, linkPanelData) => {
+const goToLink = (event, linkData, linkPanelData, experiments) => {
   const { anchor, url, target } = linkData;
   if (anchor) {
     const { customAnchorScroll } = linkPanelData;
@@ -546,7 +553,7 @@ const goToLink = (event, linkData, linkPanelData) => {
       // eslint-disable-next-line prefer-spread
       const arrayOfAllblocks = Array.apply(null, nodeListOfAllblocks);
       const element = arrayOfAllblocks.find(block => block.dataset.offsetKey === `${anchor}-0-0`);
-      anchorScroll(element);
+      anchorScroll(element, experiments);
     }
   } else {
     const href = url ? normalizeUrl(url) : undefined;